refactor(shard): Phase 3 P0/P1 插件瘦身 — duel/dream/repeater
将 repeater 全模块、duel 八角笼/QTE、dream 漂流与历史梦库迁到 shard.context 薄适配层，行为不变。
