+++ src/plugins/repeater/message_store.py
@@ -9,6 +9,7 @@ from nonebot import logger
 
 from src.foundation.db import Message as MessageModel
 from src.foundation.db import make_message_repository
+from src.platform.shard import context as shard_ctx
 
 from .config import get_repeater_config
 
@@ -86,9 +87,7 @@ class MessageStore:
         if should_sync:
             await MessageStore._sync(cur_time)
 
-        from src.platform.shard.registry.config import is_sharding_active
-
-        if is_sharding_active():
+        if shard_ctx.sharding_active():
             from src.platform.shard.coord.repeater_buffer import schedule_publish_repeater_buffer
 
             schedule_publish_repeater_buffer(chat_data)
