+++ src/plugins/repeater/reply_record_sync.py
@@ -4,11 +4,12 @@ from __future__ import annotations
 
 from typing import Any
 
+from src.platform.shard import context as shard_ctx
+from src.plugins.repeater.model import Chat
+
 
 def should_publish_reply_record(record: dict[str, Any]) -> bool:
     """占位符与空回复不参与跨片同步，减少无效 coord IO。"""
-    from src.plugins.repeater.model import Chat
-
     reply = str(record.get("reply") or "").strip()
     if not reply:
         return False
@@ -18,9 +19,7 @@ def should_publish_reply_record(record: dict[str, Any]) -> bool:
 def publish_reply_record(group_id: int, bot_id: int, record: dict[str, Any]) -> None:
     if not should_publish_reply_record(record):
         return
-    from src.platform.shard.registry.config import is_sharding_active
-
-    if not is_sharding_active():
+    if not shard_ctx.sharding_active():
         return
     from src.platform.shard.coord.repeater_reply_buffer import schedule_publish_repeater_reply_record
