+++ src/plugins/repeater/shard_opt.py
@@ -2,12 +2,12 @@
 
 from __future__ import annotations
 
+from src.platform.shard import context as shard_ctx
+
 
 def repeater_worker_handles_message(bot_id: int) -> bool:
     """群消息入口统一放开到本 worker 本地牛，由后续去重/claim 收敛到单牛或 fanout。"""
-    from src.platform.shard.registry.config import is_sharding_active
-
-    if not is_sharding_active():
+    if not shard_ctx.sharding_active():
         return True
     return True
 
@@ -23,19 +23,13 @@ def local_connected_bot_ids() -> frozenset[int]:
 
 def repeater_scheduler_runs_on_worker() -> bool:
     """主动发言定时任务：分片时仅代表牛所在 worker 执行，减少重复扫描。"""
-    from src.platform.shard.registry.config import is_sharding_active
-
-    if not is_sharding_active():
+    if not shard_ctx.sharding_active():
         return True
-    from src.platform.shard.local_representative import local_worker_representative_bot_id
-
-    return local_worker_representative_bot_id() is not None
+    return shard_ctx.local_representative_bot_id() is not None
 
 
 def repeater_maintenance_runs_on_worker() -> bool:
     """跨 worker 全局维护（sync / context 清理）：分片时仅 shard 0 执行。"""
-    from src.platform.shard.registry.config import get_shard_registry_settings, is_sharding_active
-
-    if not is_sharding_active():
+    if not shard_ctx.sharding_active():
         return True
-    return int(get_shard_registry_settings().shard_id) == 0
+    return shard_ctx.shard_id() == 0
